Refact: Seperate createNewGroup function to reuse
