+++ src/pages/NewGroup/NewGroup.js
@@ -1,7 +1,9 @@
 import { Component } from '../../../library/CBD/index.js';
+import solver from '../../core/solver.js';
+import { getMemberIds, getMembersLength, getRecords } from '../../state/index.js';
 import MainLayout from '../../components/MainLayout/MainLayout.js';
-import Result from './Result.js';
 import SelectGroupCnt from './SelectGroupCnt.js';
+import Result from './Result.js';
 
 export default class NewGroup extends Component {
   constructor(props) {
@@ -13,6 +15,19 @@ export default class NewGroup extends Component {
     });
   }
 
+  createNewGroup(groupCnt) {
+    const data = {
+      records: getRecords(),
+      groupNum: groupCnt,
+      peopleArr: getMemberIds(),
+      totalPeopleNum: getMembersLength(),
+      forbiddenPairs: [],
+    };
+    const { newRecord } = solver(data);
+    this.props.setState({ result: newRecord, currentView: 'autoResult' });
+    console.log(newRecord);
+  }
+
   render() {
     // prettier-ignore
     return `
@@ -20,8 +35,8 @@ export default class NewGroup extends Component {
       ${new MainLayout().render()}
       <main class="main">
         ${this.resultState.currentView === 'selectGroupCnt'
-          ? new SelectGroupCnt({ setState: this.setState }).render()
-          : new Result({ resultState: this.resultState }).render()}
+          ? new SelectGroupCnt({ setState: this.setState, createNewGroup:this.createNewGroup }).render()
+          : new Result({ resultState: this.resultState, createNewGroup:this.createNewGroup }).render()}
       </main>
     </div>
     `;
